two tables, session and tmocker created with tested connection and relationship.

+++ TumamocRunKeeper/src/main/java/com/brian/springdata/TumamocRunKeeper/repos/TmockerRepository.java
@@ -2,6 +2,6 @@ package com.brian.springdata.TumamocRunKeeper.repos;
 import org.springframework.data.repository.CrudRepository;
 import com.brian.springdata.TumamocRunKeeper.entities.Tmocker;
 
-public interface TmockerRepository extends CrudRepository<Tmocker, Integer> {
+public interface TmockerRepository extends CrudRepository<Tmocker, Long> {
 
 }
